--- README.md
@@ -1,3 +1,0 @@
-# Coursework
-
-Exercises to practice and solidify your understanding of the Decomposition module of the Software Development Course.
