--- backend/package.json
@@ -1,7 +1,0 @@
-{
-  "dependencies": {
-    "cors": "^2.8.6",
-    "express": "^5.2.1",
-    "websocket": "^1.0.35"
-  }
-}
